thingsboard: fota: allow delta changes of attributes
In the case, that a firmware update has been scheduled but for some
reason did not start, e.g. because it is identical to the current
firmware, and then the target firmware is changed, thingsboard might
only send the changed attributes.

The current logic in this case dismisses the firmware update request.
Change the logic, to cache the already known values.