thingsboard: make time synchronization optional
Make the time synchronization mechanism optional. This is especially
useful, as this mechanism is not a standard thingsboard mechanism.

To avoid having to #ifdef each function out, the code handling the time
synchronization has been moved mostly untouched into 'src/tb_time.c'.

This also removes the dogy wait for time for 10 s.