thingsboard: make JSON dataformat work again
Some architectonical changes have been made with the protobuf support,
which broke the JSON capabilities. The JSON encoding and decoding has
been adapted in a similar way to make it work again. Some cleanups in
the buildsystem have been made, too, which became neccessary as the
amount of JSON schemas increased and could not handled anymore in the
same way.

Be aware, that still some things are different between on how JSON and
Protobuf APIs work now. The nanopb will generate the objects with an
static array allocated for strings, while the JSON code generator just
uses char pointers. Applications need to deal with this differences.