thingsboard: fota: move prototypes to `tb_internal.h`
In past, most headers of individual compile units of the Thingsboard SDK
have been moved into `tb_internal.h`. As some internal stuff, like the
access token, is needed by all compile units, all of them are already
including this file.

When using DTLS, the paths of the CoAP request change, as the access
token is no longer used. Removing the access token from function
prototypes, so it doesn't get in the way for the DTLS changes.