thingsboard: allow communication in the connecting state
During the `THINGSBOARD_STATE_CONNECTING` state, the socket already has
been setup, but `thingsboard_is_active()` will return `false` and some
communication, like the initial reporting of the FOTA state, is
supressed. To fix that problem, communication must be allowed during the
`THINGSBOARD_STATE_CONNECTING` state.

The naming is now a bit odd, as the "active" state of a device does not
really correspond in a strict manner with the returned value of that
function any more.